requirements.txt
streamlit == 1.49.1
python-dotenv
nltk
langchain-community
langchain-text-splitters
nltk
langchain-chroma
chromadb
langchain-huggingface
langchain-groq
sentence-transformers
unstructured[pdf]
unstructured
transformers
httpx== 0.27.0
youtube-search-python